+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/controller/v1/system/LoginParamRecordController.java
@@ -10,6 +10,7 @@ import com.dwarfeng.subgrade.sdk.bean.dto.PagingUtil;
 import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.login.LoginRequired;
+import com.dwarfeng.subgrade.sdk.interceptor.permission.PermissionRequired;
 import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
 import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
@@ -52,6 +53,7 @@ public class LoginParamRecordController {
     @GetMapping("/login-param-record/{loginHistoryId}&{recordId}/exists")
     @BehaviorAnalyse
     @LoginRequired
+    @PermissionRequired("webapi.controller_permitted.system.login_param_record.exists")
     public FastJsonResponseData<Boolean> exists(
             HttpServletRequest request,
             @PathVariable("loginHistoryId") Long loginHistoryId, @PathVariable("recordId") String recordId
@@ -68,6 +70,7 @@ public class LoginParamRecordController {
     @GetMapping("/login-param-record/{loginHistoryId}&{recordId}")
     @BehaviorAnalyse
     @LoginRequired
+    @PermissionRequired("webapi.controller_permitted.system.login_param_record.get")
     public FastJsonResponseData<JSFixedFastJsonLoginParamRecord> get(
             HttpServletRequest request,
             @PathVariable("loginHistoryId") Long loginHistoryId, @PathVariable("recordId") String recordId
@@ -88,6 +91,7 @@ public class LoginParamRecordController {
     @GetMapping("/login-param-record/all")
     @BehaviorAnalyse
     @LoginRequired
+    @PermissionRequired("webapi.controller_permitted.system.login_param_record.all")
     public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonLoginParamRecord>> all(
             HttpServletRequest request,
             @RequestParam("page") int page, @RequestParam("rows") int rows
@@ -109,6 +113,7 @@ public class LoginParamRecordController {
     })
     @BehaviorAnalyse
     @LoginRequired
+    @PermissionRequired("webapi.controller_permitted.system.login_param_record.child_for_login_history")
     public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonLoginParamRecord>> childForLoginHistory(
             HttpServletRequest request,
             @PathVariable(required = false, value = "loginHistoryId") Long loginHistoryId,